fix copilot, file split
