+++ init.lua
@@ -38,7 +38,6 @@ I hope you enjoy your Neovim journey,
 P.S. You can delete this when you're done too. It's your config now :)
 --]]
 
-vim.o.wildmenu = true
 
 -- Set <space> as the leader key
 -- See `:help mapleader`
@@ -247,10 +246,7 @@ require('lazy').setup({
 -- See `:help vim.o`
 -- NOTE: You can change these options as you wish!
 
-vim.o.tabstop = 2
-vim.o.shiftwidth = 2
-vim.o.softtabstop=2
-vim.o.expandtab = true
+require 'custom.options'
 
 -- Set highlight on search
 vim.o.hlsearch = true
@@ -289,8 +285,6 @@ vim.o.completeopt = 'menuone,noselect'
 -- NOTE: You should make sure your terminal supports this
 vim.o.termguicolors = true
 
-vim.o.list = true
-vim.o.listchars = "eol:↩,tab:<·>"
 
 -- [[ Basic Keymaps ]]
 
@@ -646,56 +640,4 @@ cmp.setup {
 -- The line beneath this is called `modeline`. See `:help modeline`
 -- vim: ts=2 sts=2 sw=2 et
 
-vim.g.carlos_colorscheme = "gruvbox"
-
-function ColorGruvBox()
-  vim.g.gruvbox_contrast_dark = 'hard'
-  vim.g.gruvbox_italic = '1'
-  vim.g.gruvbox_bold = '1'
-  vim.g.gruvbox_invert_selection = '0'
-  vim.opt.background = "dark"
-
-  vim.cmd("colorscheme " .. vim.g.carlos_colorscheme)
-
-  local hl = function(thing, opts)
-    vim.api.nvim_set_hl(0, thing, opts)
-  end
-
-  hl("SignColumn", {
-    bg = "none",
-  })
-
-  -- hl("ColorColumn", {
-  --     ctermbg = 0,
-  --     bg = "#555555",
-  -- })
-
-  -- hl("CursorLineNR", {
-  --   bg = "None"
-  -- })
-
-  hl('WinSeparator', {
-    fg = 'Gray'
-  })
-
-  hl('Cursor', {
-    bg = 'DarkCyan',
-    fg = 'White'
-  })
-
-  hl("Normal", {
-    bg = "none"
-  })
-
-  -- hl("LineNr", {
-  --     fg = "#5eacd3"
-  -- })
-
-  -- hl("netrwDir", {
-  --     fg = "#5eacd3"
-  -- })
-end
-
-ColorGruvBox()
-
-vim.g.skip_ts_context_commentstring_module = true
+require 'custom.colors'
