+++ lua/custom/options.lua
@@ -1,0 +1,13 @@
+vim.o.wildmenu = true
+vim.o.tabstop = 2
+vim.o.shiftwidth = 2
+vim.o.softtabstop = 2
+vim.o.expandtab = true
+vim.o.scrolloff = 2
+
+vim.o.list = true
+vim.o.listchars = "eol:↩,tab:<·>"
+
+vim.g.carlos_colorscheme = "gruvbox"
+
+vim.g.skip_ts_context_commentstring_module = true
